Refactor v2 parser steps
